+++ src/components/forms/forms.ts
@@ -29,31 +29,44 @@ const delayDetectionAllInputs = (inputElements: NodeListOf<Element>) => {
  */
 const detectInput = (input: any) => {
   const formField = input.closest('.form-field');
+  const customSelect = formField.querySelector('.form-custom-select');
 
   const isActive = formField.classList.contains('active');
   const types = ['date', 'month', 'week', 'time'];
-  const hasContent =
-    input.value.length > 0 ||
-    (input.tagName !== 'SELECT' && input.placeholder.length > 0) ||
-    input.tagName === 'SELECT' ||
-    types.some((type) => input.matches(`[type="${type}"]`));
+
+  let hasContent = customSelect && input.tagName === 'DIV' && input.innerText.length > 0;
+  if (!customSelect)
+    hasContent =
+      input.value.length > 0 ||
+      (input.tagName !== 'SELECT' && input.placeholder.length > 0) ||
+      input.tagName === 'SELECT' ||
+      types.some((type) => input.matches(`[type="${type}"]`));
+
   const isFocused = document.activeElement === input;
   const isDisabled = input.hasAttribute('disabled') || input.hasAttribute('readonly');
 
   if (input.firstInit) {
-    updateInput(input, isActive, hasContent, isFocused, formField);
+    updateInput(input, isActive, hasContent, isFocused, formField, customSelect);
     input.firstInit = false;
     input.isInit = true;
   } else {
-    if (!isDisabled) updateInput(input, isActive, hasContent, isFocused, formField);
+    if (!isDisabled) updateInput(input, isActive, hasContent, isFocused, formField, customSelect);
   }
 };
 
 /**
  * Update input field
  */
-const updateInput = (input: any, isActive: boolean, hasContent: boolean, isFocused: boolean, formField) => {
+const updateInput = (
+  input: any,
+  isActive: boolean,
+  hasContent: boolean,
+  isFocused: boolean,
+  formField,
+  customSelect: HTMLDivElement
+) => {
   const isTextArea = input.type === 'textarea';
+  const label = formField.querySelector('label:not(.form-check)');
 
   if (!isActive && (hasContent || isFocused)) {
     formField.classList.add('active');
@@ -61,15 +74,11 @@ const updateInput = (input: any, isActive: boolean, hasContent: boolean, isFocus
     formField.classList.remove('active');
   }
 
-  if (!isTextArea) setFormPosition(input, formField);
+  if (!isTextArea) setFormPosition(input, formField, customSelect, label);
+  else if (label) label.style.backgroundColor = getLabelColor(label);
 
   if (isFocused && !isTextArea) formField.classList.add('is-focused');
-  else formField.classList.remove('is-focused');
-
-  if (isTextArea) {
-    const label = formField.querySelector('label');
-    label.style.backgroundColor = getLabelColor(label);
-  }
+  else if (!customSelect) formField.classList.remove('is-focused');
 
   if (isFocused && isTextArea) formField.classList.add('is-textarea-focused');
   else formField.classList.remove('is-textarea-focused');
@@ -78,11 +87,16 @@ const updateInput = (input: any, isActive: boolean, hasContent: boolean, isFocus
 /**
  * Add bottom position variable to form
  */
-const setFormPosition = (input: HTMLElement, formField: HTMLElement) => {
+const setFormPosition = (
+  input: HTMLElement,
+  formField: HTMLElement,
+  customSelect: HTMLDivElement,
+  label?: HTMLLabelElement
+) => {
   const inputWidth = input.clientWidth,
     inputLeftOffset = input.offsetLeft;
 
-  const topOffset = input.clientHeight + input.offsetTop + 'px';
+  const topOffset = input.clientHeight + (customSelect ? customSelect.offsetTop : input.offsetTop) + 'px';
   const isBordered = input.closest('.form-material').classList.contains('form-material-bordered');
 
   formField.style.setProperty(getCssVar('form-material-position'), topOffset);
@@ -102,7 +116,6 @@ const setFormPosition = (input: HTMLElement, formField: HTMLElement) => {
   if (offset != 0) labelLeft = inputLeftOffset;
   formField.style.setProperty(getCssVar('form-material-width'), width);
 
-  const label = formField.querySelector('label');
   if (label) {
     label.style.left = labelLeft + 'px';
 
@@ -208,7 +221,9 @@ const updateInputsFile = () => {
  * Update inputs state
  */
 export const updateInputs = (
-  inputElements = document.querySelectorAll('.form-material .form-field:not(.form-default) .form-control')
+  inputElements: NodeListOf<HTMLElement> | Array<HTMLElement> = document.querySelectorAll(
+    '.form-material .form-field:not(.form-default) .form-control:not(.form-custom-select)'
+  )
 ) => {
   const { setupInputs, detectInputs } = Array.from(inputElements).reduce(
     (acc, el: any) => {